+++ R/smooth.R
@@ -213,3 +213,113 @@ setMethod(
                          sparse = sparse)
   }
 )
+
+# Penalized likelihood smoothing ===============================================
+#' @export
+#' @rdname smooth_likelihood
+#' @aliases smooth_likelihood,numeric,numeric-method
+setMethod(
+  f = "smooth_likelihood",
+  signature = signature(x = "numeric", y = "numeric"),
+  definition = function(x, y, lambda, d = 2, AIC = TRUE, SE = FALSE,
+                        progress = interactive()) {
+
+    if (!requireNamespace("Matrix", quietly = TRUE)) {
+      msg <- "The Matrix package is required. Please install it."
+      stop(msg, call. = FALSE)
+    }
+
+    m <- length(y)
+    n <- length(lambda)
+
+    aic <- rep(0, n)
+    mu <- matrix(data = 0, nrow = m, ncol = n)
+    se <- matrix(data = 0, nrow = m, ncol = n)
+    R <- matrix(data = 0, nrow = m, ncol = n)
+
+    if (progress) pb <- utils::txtProgressBar(min = 0, max = n, style=3)
+
+    n_lambda <- seq_len(n)
+    for (i in n_lambda) {
+      like <- penalized_likelihood(y, lambda = lambda[i], d = d,
+                                   AIC = AIC, SE = SE)
+
+      aic[i] <- like$aic
+      mu[, i] <- like$mu
+      se[, i] <- like$se
+      R[, i] <- like$r
+
+      if (progress) utils::setTxtProgressBar(pb, i)
+    }
+
+    if (progress) close(pb)
+
+    k <- which.min(aic)
+    mu <- mu[, k]
+    se <- se[, k]
+    R <- R[, k]
+    op_lambda <- lambda[k]
+    op_aic <- aic[k]
+
+    xy <- list(x = x, y = mu, aic = aic, op_aic = op_aic, op_lambda = op_lambda,
+               se = se, r = R)
+    attr(xy, "method") <- "penalized likelihood smoothing"
+    xy
+  }
+)
+
+#' @export
+#' @rdname smooth_likelihood
+#' @aliases smooth_likelihood,ANY,missing-method
+setMethod(
+  f = "smooth_likelihood",
+  signature = signature(x = "ANY", y = "missing"),
+  definition = function(x, lambda, d = 2, AIC = TRUE, SE = FALSE) {
+    xy <- grDevices::xy.coords(x)
+    methods::callGeneric(x = xy$x, y = xy$y, lambda = lambda, d = d,
+                         AIC = AIC, SE = SE)
+  }
+)
+
+penalized_likelihood <- function(y, lambda, d = 2, AIC = TRUE, SE = FALSE) {
+
+  m <- length(y)
+  z <- log(y + 1)
+  E <- Matrix::Diagonal(m)
+  D <- Matrix::diff(E, differences = d)
+  P <- lambda * Matrix::crossprod(D, D)
+  mu <- exp(z)
+
+  delta_mu <- 1
+  limit <- 1 # Stop condition to prevent infinite loop
+  while (delta_mu >= 0.0001 & limit <= 10) {
+    W <- Matrix::Diagonal(x = mu)
+    z <- Matrix::solve(W + P, y - mu + mu * z)
+    new_mu <- exp(as.numeric(z)) # Drop S4 class
+    delta_mu <- max(abs(new_mu - mu))
+    mu <- new_mu
+
+    limit <- limit + 1
+  }
+
+  ## AIC
+  aic <- 0
+  if (isTRUE(AIC)) {
+    H <- Matrix::solve(W + P) %*% W
+    ed <- sum(Matrix::diag(H))
+    dev <- 2 * sum(y * log((y + 1e-10) / mu))
+    aic <- dev + 2 * ed * m / (m - ed)
+  }
+
+  ## Standard error
+  se <- rep(0, m)
+  if (isTRUE(SE)) {
+    HH <- Matrix::tcrossprod(H %*% W, H)
+    se <- sqrt(Matrix::diag(HH))
+  }
+
+  ## Standardized residuals for counts
+  r <- (y - mu) / sqrt(mu)
+
+  list(mu = mu, aic = aic, se = se, r = r)
+}