+++ NEWS.md
@@ -1,4 +1,6 @@
 # alkahest 1.0.0.9000
+## New classes and methods
+* Add penalized likelihood smoothing (de Rooi et al. 2014).
 
 # alkahest 1.0.0